refactor: refactor coherency to use player collision instead of per hiding achor
