--- Assets/OurFiles/Scripts/Environment/HidingTeleportTrigger.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 55e6ef8ebb014a1478655fbd05a83102
